// autogpt_platform/frontend/src/app/api/__generated__/models/accuracyTrendsResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { AccuracyLatestData } from "./accuracyLatestData";
import type { AccuracyTrendsResponseAlert } from "./accuracyTrendsResponseAlert";
import type { AccuracyTrendsResponseHistoricalData } from "./accuracyTrendsResponseHistoricalData";

/**
 * Response model for accuracy trends and alerts.
 */
export interface AccuracyTrendsResponse {
  latest_data: AccuracyLatestData;
  alert: AccuracyTrendsResponseAlert;
  historical_data?: AccuracyTrendsResponseHistoricalData;
}